Add BrownoutDetected Events - device
Brownout Detection is detected during the boot sequence.  If the
previous shutdown resulted in a reboot reason that has *ocp* or *uvlo*
in it, the shutdown was due to brownout.  Mitigation Logger should have
logged the device state during the brownout.  This event metric is to
surface the logged data.

Bug: 250009365
Test: Confirm triggering of events
Signed-off-by: George Lee <geolee@google.com>
Change-Id: Ib142a4135d0a9ff617fc8355f0ba978b5c3f8afb