Improvements for BOARD_WITHOUT_RADIO.
As far as I can tell, GPS is part of the radio board on Zuma, so disabling
the radio should disable GPS as well.

The PixelLogger app has a hard dependency on the shannondm logger,
so don't install it if the radio is disabled.

This will help avoid service restarts so that we will be able
to do more accurate perf measurements.

Bug: 237601345
Change-Id: I60639d8e8fb61343ba19c2d9bea40cbeb8c6e0fd