Set the environment variable to switch the Keymint HAL service to Rust
Ref: aosp/2328294

Note:
This CL switches the Keymint HAL to the Rust implementation which can
only communicate with a Keymint TA implemented in Rust in Trusty.
go/gpar/489149 switches the Keymint TA to the Rust implementation.

Therefore, this CL is one of two CLs that needs to be synced in order
for the zuma devices to be able to boot with
Keymint Rust implementation.

Bug: 197891150
Bug: 225036046
Test: VtsAidlKeyMintTargetTest
Change-Id: I53307b3911a478d941fc2d6cc39f3d410ad050e0