Create a folder to store the calibration file
This folder was created previously by a factory calibration script. On some 2023 devices, auto calibration will replace factory calibration, so the folder won't be automatically created in the factory anymore. We create the folder as part of the device initialization, so it is available by the time we try to write a new calibration file into the folder.

Bug: 246991030
Test: test on device
Change-Id: I8548bc7f7058f435cf3df6734de33bed62b86469