Assign owner/group for cp_temp after the kernel module is loaded.
The chown operation on the cp_temp sysfs attribute needs to happen after
the cp_thermal_zone driver has been loaded by the insmod_sh service.
If not, the owner/group won't be updated and after the driver gets
loaded, cp_temp will default to root/root. Hence, we are moving this
operation from init to the post-fs-data stage.

Test: Verified owner/group for the sysfs attribute
Bug: 267485434
Change-Id: Ib463dd5f2de332164714a77bdb0302afeaad5217
Signed-off-by: Mahesh Kallelil <kallelil@google.com>