UsbGadget: Add support for UVC function
UVC is a new USB function supported in Android. This CL adds UVC as a
valid function and gives it a new pid of 0x4ee[de]. UVC function is
guarded by the property `ro.usb.uvc.enabled`. When this property is set
to false, UVC won't be considered as a valid gadget function.

Bug: 242344221

Test: Manually tested that the UVC function is successfully configured

Merged-In: I5c2040a84ee8360b16c955403478d908fe109121
Change-Id: I5c2040a84ee8360b16c955403478d908fe109121
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>